+++ message/src/main/java/org/ietf/nea/pa/attribute/util/PaAttributeValueRemediationParameterString.java
@@ -6,7 +6,7 @@ public class PaAttributeValueRemediationParameterString extends AbstractPaAttrib
   
     private final long stringLength;          // 32 bit(s) length of the string in octets
     private final String remediationString;   // variable length, UTF-8 encoded, NUL termination MUST NOT be included.
-    private final byte langCodeLength;        // 8 bit(s) length of language code in octets, 0 = language unknown
+    private final short langCodeLength;        // 8 bit(s) length of language code in octets, 0 = language unknown
     private final String langCode;            // variable length, US-ASCII string composed of a well-formed RFC 4646 [3] language tag
     
 //    //overload the superclass default vendorId, type and length value
@@ -38,7 +38,7 @@ public class PaAttributeValueRemediationParameterString extends AbstractPaAttrib
 	/**
 	 * @return the langCodeLength
 	 */
-	public byte getLangCodeLength() {
+	public short getLangCodeLength() {
 		return this.langCodeLength;
 	}
